+++ app/src/main/java/com/alexanderpodkopaev/androidacademyproject/data/entity/MovieGenre.kt
@@ -4,6 +4,6 @@ import androidx.room.Entity
 
 @Entity(primaryKeys = ["movieId", "genreId"])
 data class MovieGenre(
-    val movieId: Long,
-    val genreId: Long
+    val movieId: Int,
+    val genreId: Int
 )
